+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Option to show the full application name with the edition in Rich Presence title
 
+### Changed
+
+- Clarified "try reconnecting" setting
+
 ## [1.6.1] - 2024-11-10
 
 ### Fixed