+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.2] - 2024-12-01
+
 ### Added
 
 - Option to show the full application name with the edition in Rich Presence title
@@ -190,7 +192,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.6.1...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.6.2...HEAD
+[1.6.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.6.1...v1.6.2
 [1.6.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.5.0...v1.5.1